+++ tests/settings.py
@@ -1,0 +1,1 @@
+SECRET_KEY = ')dajq1#olz2*y&$1x0y&pd0ev-a_h2*j%ed0j5ych2^oy%*2%e'